Moves control panel tabs into their own folder
