+++ TGControlPanel/ControlPanel/ControlPanel.resx
