+++ TGControlPanel/ControlPanel/StaticPage.cs
