+++ TGControlPanel/TGControlPanel.csproj
@@ -49,10 +49,10 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ByondPage.cs">
+    <Compile Include="ControlPanel\ByondPage.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="ChatPage.cs">
+    <Compile Include="ControlPanel\ChatPage.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Login.cs">
@@ -61,10 +61,10 @@
     <Compile Include="Login.Designer.cs">
       <DependentUpon>Login.cs</DependentUpon>
     </Compile>
-    <Compile Include="ControlPanel.cs">
+    <Compile Include="ControlPanel\ControlPanel.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="ControlPanel.Designer.cs">
+    <Compile Include="ControlPanel\ControlPanel.Designer.cs">
       <DependentUpon>ControlPanel.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
@@ -74,20 +74,20 @@
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
       <DependentUpon>Settings.settings</DependentUpon>
     </Compile>
-    <Compile Include="RepoPage.cs">
+    <Compile Include="ControlPanel\RepoPage.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="ServerPage.cs">
+    <Compile Include="ControlPanel\ServerPage.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="..\Version.cs" />
-    <Compile Include="StaticPage.cs">
+    <Compile Include="ControlPanel\StaticPage.cs">
       <SubType>Form</SubType>
     </Compile>
     <EmbeddedResource Include="Login.resx">
       <DependentUpon>Login.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="ControlPanel.resx">
+    <EmbeddedResource Include="ControlPanel\ControlPanel.resx">
       <DependentUpon>ControlPanel.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
